requirements.txt
streamlit
google-generativeai
python-dotenv
google-cloud-bigquery